refactor: 주사위 로직 파일 분리
